change: search first at local cache subfolder

+++ R/pgx-files.R
@@ -139,6 +139,8 @@ pgx.saveMatrixH5 <- function(X, h5.file, chunk = NULL) {
       index = list(1:nrow(X), 1:ncol(X))
     )
   }
+  if(is.null(rownames(X))) rownames(X) <- 1:nrow(X)
+  if(is.null(colnames(X))) colnames(X) <- paste0("V",1:ncol(X)  )
   rhdf5::h5write(rownames(X), h5.file, "data/rownames")
   rhdf5::h5write(colnames(X), h5.file, "data/colnames")
 
@@ -187,13 +189,17 @@ pgx.readOptions <- function(file = "./OPTIONS") {
 
 #' @export
 cached.csv <- function(file, FUN = read.csv, force = FALSE, ...) {
-  file2 <- paste0("cache-", gsub("[-._/]", "", file), ".rds")
+  file2 <- sub(getwd(),"",file)
+  file2 <- paste0("cache-", gsub("[-._/]", "", file2), ".rds")
   file2
   ## cache.file <- file.path("/tmp", file2)
-  cache.file <- file.path(tempdir(), file2)
+  cache.dir <- ifelse(dir.exists("cache"), "./cache", tempdir())
+  cache.file <- file.path(cache.dir, file2)
   if (force || !file.exists(cache.file)) {
+    if(!file.exists(file)) return(NULL)
     message("[cached.csv] reading from file ", file)
     csv <- FUN(file, ...)
+    message("[cached.csv] saving cache at ", cache.file)    
     saveRDS(csv, file = cache.file)
   } else {
     message("[cached.csv] reading from cache ", cache.file)
@@ -204,13 +210,17 @@ cached.csv <- function(file, FUN = read.csv, force = FALSE, ...) {
 
 #' @export
 cached.csv.s3 <- function(file, bucket, FUN = read.csv, force = FALSE, ...) {
-  file2 <- paste0("s3cache-", gsub("[-._/]", "", file), ".rds")
+  file2 <- sub(getwd(),"",file)
+  file2 <- paste0("cache-", gsub("[-._/]", "", file2), ".rds")
   file2
-  cache.file <- file.path("/tmp", file2)
+  cache.dir <- ifelse(dir.exists("cache"), "./cache", tempdir())
+  cache.file <- file.path(cache.dir, file2)
   if (force || !file.exists(cache.file)) {
+    if(!file.exists(file)) return(NULL)    
     message("[cached.csv.s3] reading from S3 bucket=", bucket, "   file=", file)
     obj <- aws.s3::get_object(object = file, bucket = bucket)
     csv <- data.table::fread(rawToChar(obj))
+    message("[cached.csv] saving cache at ", cache.file)        
     saveRDS(csv, file = cache.file)
   } else {
     message("[cached.csv.s3] reading from cache ", cache.file)
